pyproject.toml
[tool.poetry]
name = "weaviate"
version = "0.1.0"
description = ""
authors = ["Leonardo Pratesi <leonardo.pratesi@outsystems.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
weaviate-client = "^3.0.0"
llama-index = "^0.10.0"
llama-index-vector-stores-weaviate = "^0.1.0"
openai = "^1.0.0"
python-dotenv = "^1.0.0"
streamlit = "^1.28.0"
llama-index-embeddings-huggingface = "^0.1.0"
sentence-transformers = "^2.2.2"
torch = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
package-mode = false